+++ paq8/paq8h/paq8h.cpp
@@ -1394,10 +1394,10 @@ inline int mix2( Mixer &m, int s, StateMap &sm ) {
   int p0 = 255 - p1;
   m.add( p1 - p0 );
   m.add( st * static_cast<int>( static_cast<int>( static_cast<int>( n0 ) == 0 - static_cast<int>( n1 ) ) == 0 ) );
-  m.add( ( ( p1 & ( static_cast<int>( -static_cast<int>( n0 ) ) == 0 ) ) )
-         - ( ( p0 & ( static_cast<int>( -static_cast<int>( n1 ) ) == 0 ) ) ) );
-  m.add( ( ( p1 & ( static_cast<int>( -static_cast<int>( n1 ) ) == 0 ) ) )
-         - ( ( p0 & ( static_cast<int>( -static_cast<int>( n0 ) ) == 0 ) ) ) );
+  m.add( ( ( p1 & static_cast<int>( static_cast<int>( -static_cast<int>( n0 ) ) == 0 ) ) )
+         - ( ( p0 & static_cast<int>( static_cast<int>( -static_cast<int>( n1 ) ) == 0 ) ) ) );
+  m.add( ( ( p1 & static_cast<int>( static_cast<int>( -static_cast<int>( n1 ) ) == 0 ) ) )
+         - ( ( p0 & static_cast<int>( static_cast<int>( -static_cast<int>( n0 ) ) == 0 ) ) ) );
   return static_cast<int>( s > 0 );
 }
 
@@ -2353,7 +2353,7 @@ int jpegModel( Mixer &m ) {
   jassert( coef >= 0 && coef < 256 );
   const int zu = zzu[mcupos & 63], zv = zzv[mcupos & 63];
   if( hbcount == 0 ) {
-    const int mpos = ( mcupos >> 4 | ( static_cast<int>( static_cast<int>( ( mcupos & -64 ) ) ) == 0 << 7 ) );
+    const int mpos = ( mcupos >> 4 | static_cast<int>( static_cast<int>( static_cast<int>( ( mcupos & -64 ) ) ) == 0 << 7 ) );
     int n = 0;
     cxt[0] = hash( ++n, hc, mcupos >> 2, min( 3, mcupos & 63 ) );
     cxt[1] = hash( ++n, hc, mpos >> 4, cbuf[cpos - mcusize] );
